pyproject.toml
[tool.poetry]
name = "drone-detection-mlops"
version = "0.1.0"
description = "MLOps pipeline for YOLO drone detection"
authors = ["Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
ultralytics = "^8.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
gradio = "^5.0.0"
mlflow = "^2.8.0"
boto3 = "^1.29.0"
minio = "^7.2.0"
pillow = "^10.0.0"
python-multipart = ">=0.0.9"
requests = "^2.31.0"
pyyaml = "^6.0"
prometheus-client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.25.0"
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
